# ARROW-10664: [Rust] Implement AVX512 sort