Before:
```
sort 2^10               time:   [94.137 us 94.154 us 94.174 us]                      
Found 2 outliers among 100 measurements (2.00%)
  2 (2.00%) high mild

sort 2^12               time:   [483.03 us 483.08 us 483.13 us]                      
Found 8 outliers among 100 measurements (8.00%)
  1 (1.00%) low severe
  2 (2.00%) low mild
  4 (4.00%) high mild
  1 (1.00%) high severe

sort nulls 2^10         time:   [59.782 us 59.800 us 59.818 us]                            
Found 1 outliers among 100 measurements (1.00%)
  1 (1.00%) high mild

sort nulls 2^12         time:   [296.84 us 296.89 us 296.97 us]                            
Found 4 outliers among 100 measurements (4.00%)
  3 (3.00%) high mild
  1 (1.00%) high severe

```

After:
```
sort 2^10               time:   [73.098 us 73.119 us 73.148 us]                      
                        change: [-22.404% -22.381% -22.356%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 8 outliers among 100 measurements (8.00%)
  3 (3.00%) low mild
  1 (1.00%) high mild
  4 (4.00%) high severe

sort 2^12               time:   [354.82 us 354.98 us 355.23 us]                      
                        change: [-26.550% -26.530% -26.505%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 8 outliers among 100 measurements (8.00%)
  1 (1.00%) low mild
  3 (3.00%) high mild
  4 (4.00%) high severe

sort nulls 2^10         time:   [53.570 us 53.577 us 53.585 us]                            
                        change: [-10.407% -10.378% -10.350%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 6 outliers among 100 measurements (6.00%)
  3 (3.00%) low mild
  3 (3.00%) high mild

sort nulls 2^12         time:   [238.46 us 238.50 us 238.55 us]                            
                        change: [-19.670% -19.650% -19.627%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 5 outliers among 100 measurements (5.00%)
  1 (1.00%) low severe
  2 (2.00%) high mild
  2 (2.00%) high severe

```

~Note: Waiting for other bit or PR to merge to rebase on top of it.~ Merged.